iortcw-git: Add build() function - fixes #12
This fixes building with tools that break under fakeroot, such as
ccache.

ccache ends up fooled to use `/run/user/0/ccache-tmp` - a directory it
has no access to. The reason is that makepkg automatically uses
fakeroot to run commands under package(), to fool any installation
tools that may be so foolish as to require root privileges to install
into a non-root directory (`$pkgdir`).

This probably fixes other things too, but I wouldn't know.